font-awesome-bootstrap.css
The file belongs to the css/ directory. And you have to download the font-awesome glyphicons from here:
https://fortawesome.github.io/Font-Awesome/
The advantage of the extension is, that you can use the icons by the same syntax as given in bootstrap. Another advantage is, that if you change the basic glyphicons class in css, maybe the size, you only have to change one class and don't have the basic class fa which was given in the css of the font-awesome.css